requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
undetected-chromedriver==3.5.5
selenium==4.18.1 